Syntax change for implementing interfaces.
A possible future extension would be generics, but the previous
syntax for implementing interfaces was similar to the syntax for
generics used in other high level programming languages (java, c#,
vala). So now the implemented interfaces are just a comma
separated list, also starting with a comma:

    class Example : Object, IExample1, IExample2 {...}

NOTE: This is a backward incompatible change, and existing carbon
sources must be updated where applicable. Until we reach version
1.0 changes like this can happen.